Extract EPIS number and date parsing from the CSV parser
Pure move: EpisNumbers owns decimal-convention number parsing with its
grouping patterns and EpisDates the dotted-date extraction, each a
single connected unit, while EpisCsvParser keeps structural parsing and
row-value lookup; the report parsers' static imports rewire directly.
MandateDeadlines stays whole by decision - its many methods share one
period-calendar core, so PMD's flag is a false positive there.

God classes 9 -> 8, PMD violations 142 -> 140.
