+++ metrics/scorecard.json
@@ -8,7 +8,7 @@
   "disabledTests": 5,
   "disconnectedClasses": 454,
   "fqnViolations": 0,
-  "godClasses": 9,
+  "godClasses": 8,
   "interactionVerifyingTestClasses": 361,
   "interactionVerifyingTestShare": 36.2,
   "lineCoverage": 84.03,
@@ -26,17 +26,17 @@
   "overMockedTestClasses": 111,
   "overMockedTestShare": 11.1,
   "pitestSlicedModules": 52,
-  "pmdViolations": 142,
+  "pmdViolations": 140,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 35,
-    "CyclomaticComplexity": 74,
-    "GodClass": 9,
+    "CyclomaticComplexity": 73,
+    "GodClass": 8,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 118.9,
+  "slowest10TestClassesSeconds": 116.2,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 200,
+  "suiteWallTimeSeconds": 221,
   "testClassesOver10Seconds": 5,
   "testClassesTotal": 998,
   "testIntegrationShare": 6.7,
