+++ src/main/java/ee/tuleva/onboarding/investment/epis/parser/R16ReportParser.java
@@ -1,7 +1,7 @@
 package ee.tuleva.onboarding.investment.epis.parser;
 
 import static ee.tuleva.onboarding.investment.epis.parser.EpisCsvParser.findValue;
-import static ee.tuleva.onboarding.investment.epis.parser.EpisCsvParser.parseNumber;
+import static ee.tuleva.onboarding.investment.epis.parser.EpisNumbers.parseNumber;
 
 import ee.tuleva.onboarding.fund.TulevaFund;
 import java.math.BigDecimal;
